@@ -1,40 +1,32 @@
-// swift-tools-version: 5.9
+// swift-tools-version: 6.2
 import PackageDescription
 
-let package = Package(
-    name: "Hyperprompt",
-    platforms: [
-        .macOS(.v12)
-    ],
-    products: [
-        .executable(
-            name: "hyperprompt",
-            targets: ["CLI"]
-        ),
-        .library(
-            name: "HypercodeGrammar",
-            targets: ["HypercodeGrammar"]
-        ),
-        .library(
-            name: "EditorEngine",
-            targets: ["EditorEngine"]
-        ),
-    ],
-    dependencies: [
-        .package(
-            url: "https://github.com/apple/swift-argument-parser",
-            from: "1.2.0"
-        ),
-        .package(
-            url: "https://github.com/apple/swift-crypto",
-            from: "3.0.0"
-        ),
-        .package(
-            url: "https://github.com/SoundBlaster/SpecificationCore",
-            from: "1.0.0"
-        ),
-    ],
-    targets: [
+let editorTrait = Trait(
+    name: "Editor",
+    description: "Enable the EditorEngine module",
+    enabledTraits: []
+)
+
+let enabledTraits = Set(
+    (Context.environment["SWIFT_PACKAGE_TRAITS"] ?? "")
+        .split(separator: ",")
+        .map { String($0) }
+)
+let enableAllTraits = Context.environment["SWIFT_ENABLE_ALL_TRAITS"] == "1"
+let editorEnabled = enableAllTraits || enabledTraits.contains(editorTrait.name)
+
+var products: [Product] = [
+    .executable(
+        name: "hyperprompt",
+        targets: ["CLI"]
+    ),
+    .library(
+        name: "HypercodeGrammar",
+        targets: ["HypercodeGrammar"]
+    ),
+]
+
+var targets: [Target] = [
         // Core module
         .target(
             name: "Core",
